requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
selenium==4.15.2
google-generativeai==0.3.2
webdriver-manager==4.0.1
